26/09

+++ AD-Ud1-AT.02-File_Iglesias_Fernandez_Xabier/src/ex2/ClasificaDirectorio.java
@@ -29,24 +29,27 @@ public class ClasificaDirectorio {
                 //Creacion directorios según las extensiones
                 for (int i = 0; i<extensiones.length; i++) {
                     if(extensiones[i] != null){
-                        File newDirectory = new File(extensiones[i].toUpperCase());
-                        
+
+                        File newDirectory = new File(archivo, extensiones[i].toUpperCase());
+
                         if(!newDirectory.exists()){
                             newDirectory.mkdir();
                         }
 
                         //Mover archivos a los nuevos directorios
                         
-                        File destino = new File(newDirectory, archivos[i].getAbsolutePath());
+                        File destino = new File(newDirectory, archivos[i].getName());
 
                         String extension = "";
                         if(archivos[i].getName().split(regex)[1]!=null){
                             extension = archivos[i].getName().split(regex)[1];
                             System.out.println(extension.toUpperCase());
                         }
-                        if(extension.equals(extensiones[i]) && newDirectory.exists()){
-                            boolean move = archivos[i].renameTo(destino);
-                            System.out.println(move);
+                        if(archivos[i].isFile()){
+                            if(extension.equals(extensiones[i]) && newDirectory.exists()){
+                                boolean move = archivos[i].renameTo(destino);
+                                System.out.println(move);
+                            }
                         }
                     }
                 }
@@ -54,7 +57,5 @@ public class ClasificaDirectorio {
         }else{
             System.out.println("El directorio indicado no existe.");
         }
-
     }
-
 }
